Some warnings (Mac) eliminated.

+++ generic/tkTextDisp.c
@@ -12173,7 +12173,7 @@ TkTextGetDLineInfo(
     *heightPtr = dlPtr->height;
 
     if (extents) {
-	*widthPtr = MAX(0, *xPtr + dlPtr->length - dInfoPtr->maxX);
+	*widthPtr = MAX(0, *xPtr + (int) dlPtr->length - dInfoPtr->maxX);
 	*heightPtr = MAX(0, *yPtr + *heightPtr - dInfoPtr->maxY);
 	*xPtr = MAX(0, -*xPtr);
 	*yPtr = MAX(0, -*yPtr);
